+++ src/components/Tab/TabsWrapper.js
@@ -1,13 +1,17 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
+import cx from 'classnames';
 import styles from './style.scss';
 
-const TabsWrapper = ({ children }) => (
-  <div className={styles.tabs}>{children}</div>
+const TabsWrapper = ({ children, className, ...restProps }) => (
+  <div {...restProps} className={cx(styles.tabs, { [className]: className })}>
+    {children}
+  </div>
 );
 
 TabsWrapper.propTypes = {
-  children: PropTypes.node.isRequired
+  children: PropTypes.node.isRequired,
+  className: PropTypes.string
 };
 
 export default TabsWrapper;